[Module] Implemented workaround for fileShare role assignment (#4391)
* Update to latest

* Update to latest

+++ utilities/pipelines/staticValidation/helper/helper.psm1
@@ -161,14 +161,14 @@ function Remove-JSONMetadata {
         # Case: Hashtable
         $resourceIdentifiers = $TemplateObject.resources.Keys
         for ($index = 0; $index -lt $resourceIdentifiers.Count; $index++) {
-            if ($TemplateObject.resources[$resourceIdentifiers[$index]].type -eq 'Microsoft.Resources/deployments') {
+            if ($TemplateObject.resources[$resourceIdentifiers[$index]].type -eq 'Microsoft.Resources/deployments' -and $TemplateObject.resources[$resourceIdentifiers[$index]].properties.template.GetType().BaseType.Name -eq 'Hashtable') {
                 $TemplateObject.resources[$resourceIdentifiers[$index]] = Remove-JSONMetadata -TemplateObject $TemplateObject.resources[$resourceIdentifiers[$index]].properties.template
             }
         }
     } else {
         # Case: Array
         for ($index = 0; $index -lt $TemplateObject.resources.Count; $index++) {
-            if ($TemplateObject.resources[$index].type -eq 'Microsoft.Resources/deployments') {
+            if ($TemplateObject.resources[$index].type -eq 'Microsoft.Resources/deployments' -and $TemplateObject.resources[$index].properties.template.GetType().BaseType.Name -eq 'Hashtable') {
                 $TemplateObject.resources[$index] = Remove-JSONMetadata -TemplateObject $TemplateObject.resources[$index].properties.template
             }
         }
